requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.0
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.2.0
flask-cors==4.0.0
PyJWT==2.8.0